fix(blink): don't allow double initialization
