+++ extensions/warp-blink-wrtc/src/signaling.rs
@@ -34,3 +34,28 @@ pub enum InitiationSignal {
     #[display(fmt = "Offer")]
     Offer { call_info: CallInfo },
 }
+
+pub mod ipfs_routes {
+    use uuid::Uuid;
+    use warp::crypto::DID;
+
+    const TELECON_BROADCAST: &str = "telecon";
+    const OFFER_CALL: &str = "offer_call";
+    /// subscribe/unsubscribe per-call
+    /// CallSignal
+    pub fn call_signal_route(call_id: &Uuid) -> String {
+        format!("{TELECON_BROADCAST}/{call_id}")
+    }
+
+    /// subscribe/unsubscribe per-call
+    /// PeerSignal
+    pub fn peer_signal_route(peer: &DID, call_id: &Uuid) -> String {
+        format!("{TELECON_BROADCAST}/{call_id}/{peer}")
+    }
+
+    /// subscribe to this when initializing Blink
+    /// InitiationSignal
+    pub fn call_initiation_route(peer: &DID) -> String {
+        format!("{OFFER_CALL}/{peer}")
+    }
+}
