+++ tools/blink-webrtc-cli/src/main.rs
@@ -12,8 +12,6 @@ use warp::blink::{
     MimeType, VideoCodec,
 };
 
-use warp_blink_wrtc::BlinkImpl;
-
 use std::path::Path;
 
 use std::str::FromStr;
@@ -304,7 +302,7 @@ async fn main() -> anyhow::Result<()> {
         }
     };
 
-    let mut blink: Box<dyn Blink> = Box::new(BlinkImpl::new(multipass).await?);
+    let mut blink: Box<dyn Blink> = warp_blink_wrtc::init(multipass).await?;
     let event_stream = blink.get_event_stream().await?;
     let handle = tokio::spawn(async {
         if let Err(e) = handle_event_stream(event_stream).await {
